Remove requirement for nonempty Qalpha in Branching Conj moves
The background is that one would need information about standard Conjunctions in the induciton step. But this is not available in a proper structural induction on the hml_srbb datatype as it is defined.

Hope is that a proof on a game simplification could be used to reestablish the smaller game.